adding angle to controlling the motor
